Update coroutines to 1.6, use MainCoroutineRule
Use https://github.com/android/architecture-samples/blob/main/app/src/sharedTest/java/com/example/android/architecture/blueprints/todoapp/MainCoroutineRule.kt

update tests according to https://github.com/Kotlin/kotlinx.coroutines/blob/master/kotlinx-coroutines-test/MIGRATION.md
