+++ coroutines-codelab/build.gradle
@@ -27,7 +27,7 @@ buildscript {
         def arch_version = '2.1.0'
         def appcompat_version = '1.4.1'
         def constraint_layout_version = '2.1.3'
-        def coroutines_android_version = '1.5.2'
+        def coroutines_android_version = '1.6.0'
         def espresso_version = '3.4.0'
         def gson_version = '2.9.0'
         def junit_version = '4.13.2'
