+++ coroutines-codelab/start/src/test/java/com/example/android/kotlincoroutines/main/MainViewModelTest.kt
@@ -19,14 +19,14 @@ package com.example.android.kotlincoroutines.main
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.example.android.kotlincoroutines.fakes.MainNetworkFake
 import com.example.android.kotlincoroutines.fakes.TitleDaoFake
-import com.example.android.kotlincoroutines.main.utils.MainCoroutineScopeRule
+import com.example.android.kotlincoroutines.main.utils.MainCoroutineRule
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 
 class MainViewModelTest {
     @get:Rule
-    val coroutineScope = MainCoroutineScopeRule()
+    val coroutineScope = MainCoroutineRule()
     @get:Rule
     val instantTaskExecutorRule = InstantTaskExecutorRule()
